first android interaction with servlet

+++ src/main/java/Snapshot.java
@@ -140,7 +140,7 @@ public class Snapshot extends HttpServlet {
             OutputStreamWriter writer = new OutputStreamWriter(response.getOutputStream());
             
             String snapshotTitle = snapshotRequest.get(0).toString();
-            int snapshotType = snapshotRequest.get(1);
+            int snapshotType = (Integer)snapshotRequest.get(1);
             int[] yearColors = (int[])snapshotRequest.get(2);
             
             String reply = "Request recieved! Title will be " + snapshotTitle + ". The snapshot will be of type " + snapshotType + ".";
